src/requirements.txt
channels>=2.4.0,<3.0
channels-redis>=3.1.0,<4.0
Django>=3.0,<4.0
djangorestframework>=3.0,<4.0
djoser>=2.0,<3.0
django-cors-headers>=3.5,<4.0
psycopg2>=2.8,<3.0